move quest/models to qe_models
