+++ qe_models/__init__.py
@@ -33,7 +33,7 @@ import importlib
 # QE MODEL FACTORY-like
 def get(model_name, params):
     try:
-        qe_model = getattr(importlib.import_module('quest.models.{}'.format(QE_MODELS[model_name.lower()])), model_name)
+        qe_model = getattr(importlib.import_module('qe_models.{}'.format(QE_MODELS[model_name.lower()])), model_name)
         return qe_model(params)
 
     except ValueError as e:
