+++ qe_models/birnn_doc_att.py
