+++ qe_models/birnn_word.py
